# api/openapi.yaml
openapi: 3.0.3
info:
  title: UpNext API
  version: "0.1.0"
  description: |
    Minimal spec for the current POC endpoints. Lists artists and events from MongoDB.
servers:
  - url: http://localhost:8080
paths:
...
  /api/artists:
    get:
      summary: List artists
      operationId: listArtists
      responses:
        '200':
          description: A list of artists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Artist'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/events:
    get:
      summary: List events
      operationId: listEvents
      responses:
        '200':
          description: A list of events
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Event'
        '500':
          description: Server error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...
  /api/lists:
    get:
      summary: List lists
      operationId: listLists
      responses:
        '200':
          description: Lists
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/List'
    post:
      summary: Create list
      operationId: createList
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name:
                  type: string
              required: [name]
      responses:
        '201':
          description: Created
...
components:
  schemas:
    Artist:
      type: object
      properties:
        id:
          type: string
          description: Document id
          example: 64f1a2b3c4d5e6f7a8b9c0d1
        name:
          type: string
        synopsis:
          $ref: '#/components/schemas/Synopsis'
      required: [id, name]
    Synopsis:
      type: object
      properties:
        text:
          type: string
    Event:
      type: object
      properties:
        id:
          type: string
          description: Document id
          example: 64f1a2b3c4d5e6f7a8b9c0d1
        title:
          type: string
        startsAt:
          type: string
          format: date-time
        tourName:
          type: string
      required: [id, title]
    Error:
      type: object
      properties:
        error:
          type: string
    List:
      type: object
      properties:
        id:
          type: string
        name:
          type: string
        key:
          type: string
        isSystem:
          type: boolean
...